references/ADVANCED-transcription/speech-to-text/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
datasets==2.16.1
decorator==5.1.1
dill==0.3.7
docopt==0.6.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.20.2
idna==3.6
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
numba==0.58.1
numpy==1.26.3
packaging==23.2
pandas==2.1.4
platformdirs==4.1.0
pooch==1.8.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pycparser==2.21
pydub==0.25.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
webvtt-py==0.4.6
xxhash==3.4.1
yarl==1.9.4
